requirements.txt
streamlit
pillow
pytesseract
pdf2image
tensorflow
numpy